+++ src-tests/Console/Command/RunCommandTest.php
@@ -2,6 +2,8 @@
 
 namespace Lmc\Steward\Console\Command;
 
+use Assert\InvalidArgumentException;
+use Lmc\Steward\Console\Application;
 use Lmc\Steward\Console\CommandEvents;
 use Lmc\Steward\Console\Event\BasicConsoleEvent;
 use Lmc\Steward\Console\Event\ExtendedConsoleEvent;
@@ -9,7 +11,6 @@ use Lmc\Steward\Process\ProcessSet;
 use Lmc\Steward\Process\ProcessSetCreator;
 use Lmc\Steward\Selenium\SeleniumServerAdapter;
 use PHPUnit\Framework\TestCase;
-use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Tester\CommandTester;
 use Symfony\Component\EventDispatcher\EventDispatcher;
@@ -80,7 +81,7 @@ class RunCommandTest extends TestCase
             $errorBeginning,
             $directoryOption
         );
-        $this->expectException(\RuntimeException::class);
+        $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage($expectedError);
 
         $this->tester->execute(
@@ -278,7 +279,7 @@ class RunCommandTest extends TestCase
         $this->assertSame(1, $this->tester->getStatusCode());
     }
 
-    public function testShouldDispatchConfigureEvent()
+    public function testShouldDispatchEventsOnExecute()
     {
         $dispatcherMock = $this->getMockBuilder(EventDispatcher::class)
             ->setMethods(['dispatch'])
@@ -288,29 +289,11 @@ class RunCommandTest extends TestCase
             ->method('dispatch')
             ->with($this->equalTo(CommandEvents::CONFIGURE), $this->isInstanceOf(BasicConsoleEvent::class));
 
-        $application = new Application();
-        $application->add(new RunCommand($dispatcherMock));
-        /** @var RunCommand $command */
-        $command = $application->find('run');
-        $command->setSeleniumAdapter($this->getSeleniumAdapterMock());
-
-        (new CommandTester($command))->execute(
-            [
-                'command' => $command->getName(),
-                'environment' => 'staging',
-                'browser' => 'firefox',
-                '--tests-dir' => __DIR__ . '/Fixtures/tests',
-            ]
-        );
-    }
-
-    public function testShouldDispatchInitEvent()
-    {
-        $dispatcherMock = $this->getMockBuilder(EventDispatcher::class)
-            ->setMethods(['dispatch'])
-            ->getMock();
-
         $dispatcherMock->expects($this->at(1))
+            ->method('dispatch')
+            ->with($this->equalTo(CommandEvents::PRE_INITIALIZE), $this->isInstanceOf(ExtendedConsoleEvent::class));
+
+        $dispatcherMock->expects($this->at(2))
             ->method('dispatch')
             ->with($this->equalTo(CommandEvents::RUN_TESTS_INIT), $this->isInstanceOf(ExtendedConsoleEvent::class));