+++ src-tests/Console/Configuration/Fixtures/DoesNotImplementInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Lmc\Steward\Console\Configuration\Fixtures;
+
+class DoesNotImplementInterface
+{
+}