+++ src-tests/Selenium/CapabilitiesResolverTest.php
@@ -6,6 +6,8 @@ use Facebook\WebDriver\Remote\DesiredCapabilities;
 use Facebook\WebDriver\Remote\WebDriverBrowserType;
 use Facebook\WebDriver\WebDriverPlatform;
 use Lmc\Steward\ConfigProvider;
+use Lmc\Steward\ConfigProviderHelper;
+use Lmc\Steward\Selenium\Fixtures\CapabilitiesResolverFixture;
 use Lmc\Steward\Test\AbstractTestCase;
 use OndraM\CiDetector\Ci\Jenkins;
 use OndraM\CiDetector\CiDetector;
@@ -25,28 +27,21 @@ class CapabilitiesResolverTest extends TestCase
         /** @var AbstractTestCase $test */
         $test = $this->getMockForAbstractClass(AbstractTestCase::class, ['name'], 'FooBarTest');
 
-        $configMock = $this->createMock(ConfigProvider::class);
-        $configMock->expects($this->any())
-            ->method('__get')
-            ->willReturnMap(
-                [
-                    ['browserName', $browser],
-                ]
-            );
+        $configMock = new ConfigProviderHelper(['browserName' => $browser]);
 
         $resolver = new CapabilitiesResolver($configMock);
         $resolver->setCiDetector($this->createConfiguredMock(CiDetector::class, ['isCiDetected' => false]));
 
-        $requiredCapabilities = $resolver->resolveDesiredCapabilities($test);
-        $this->assertInstanceOf(DesiredCapabilities::class, $requiredCapabilities);
+        $desiredCapabilities = $resolver->resolveDesiredCapabilities($test);
+        $this->assertInstanceOf(DesiredCapabilities::class, $desiredCapabilities);
 
-        $requiredCapabilitiesArray = $requiredCapabilities->toArray();
-        $this->assertSame($browser, $requiredCapabilitiesArray['browserName']);
-        $this->assertSame(WebDriverPlatform::ANY, $requiredCapabilitiesArray['platform']);
-        $this->assertSame('FooBarTest::name', $requiredCapabilitiesArray['name']);
+        $desiredCapabilitiesArray = $desiredCapabilities->toArray();
+        $this->assertSame($browser, $desiredCapabilitiesArray['browserName']);
+        $this->assertSame(WebDriverPlatform::ANY, $desiredCapabilitiesArray['platform']);
+        $this->assertSame('FooBarTest::name', $desiredCapabilitiesArray['name']);
 
         if ($extraCallback !== null) {
-            $extraCallback($requiredCapabilitiesArray);
+            $extraCallback($desiredCapabilitiesArray);
         }
     }
 
@@ -80,15 +75,12 @@ class CapabilitiesResolverTest extends TestCase
         /** @var AbstractTestCase $test */
         $test = $this->getMockForAbstractClass(AbstractTestCase::class, ['name'], 'FooBarTest');
 
-        $configMock = $this->createMock(ConfigProvider::class);
-        $configMock->expects($this->any())
-            ->method('__get')
-            ->willReturnMap(
-                [
-                    ['browserName', WebDriverBrowserType::FIREFOX],
-                    ['env', 'staging'],
-                ]
-            );
+        $configMock = new ConfigProviderHelper(
+            [
+                'browserName' => WebDriverBrowserType::FIREFOX,
+                'env' => 'staging',
+            ]
+        );
 
         $ciMock = $this->createConfiguredMock(Jenkins::class, ['getBuildNumber' => 1337, 'getCiName' => 'Jenkins']);
         $ciDetectorMock = $this->createConfiguredMock(
@@ -99,10 +91,10 @@ class CapabilitiesResolverTest extends TestCase
         $resolver = new CapabilitiesResolver($configMock);
         $resolver->setCiDetector($ciDetectorMock);
 
-        $requiredCapabilities = $resolver->resolveDesiredCapabilities($test);
-        $requiredCapabilitiesArray = $requiredCapabilities->toArray();
-        $this->assertSame('staging-1337', $requiredCapabilitiesArray['build']);
-        $this->assertEquals(['staging', 'Jenkins', 'FooBarTest'], $requiredCapabilitiesArray['tags']);
+        $desiredCapabilities = $resolver->resolveDesiredCapabilities($test);
+        $desiredCapabilitiesArray = $desiredCapabilities->toArray();
+        $this->assertSame('staging-1337', $desiredCapabilitiesArray['build']);
+        $this->assertEquals(['staging', 'Jenkins', 'FooBarTest'], $desiredCapabilitiesArray['tags']);
     }
 
     public function testShouldResolveRequiredCapabilities()
@@ -118,5 +110,32 @@ class CapabilitiesResolverTest extends TestCase
         // TODO: should be instance of WebDriverCapabilities interface in next php-webdriver major release
         $this->assertInstanceOf(DesiredCapabilities::class, $requiredCapabilities);
         $this->assertSame([], $requiredCapabilities->toArray());
+    }
+
+    public function testShouldCallCustomCapabilitiesResolverIfDefined()
+    {
+        /** @var AbstractTestCase $test */
+        $test = $this->getMockForAbstractClass(AbstractTestCase::class, ['name'], 'FooBarTest');
+
+        $configMock = new ConfigProviderHelper(
+            [
+                'browserName' => 'firefox',
+                'capabilitiesResolver' => CapabilitiesResolverFixture::class,
+            ]
+        );
+
+        $resolver = new CapabilitiesResolver($configMock);
+        $resolver->setCiDetector($this->createConfiguredMock(CiDetector::class, ['isCiDetected' => false]));
+
+        $desiredCapabilities = $resolver->resolveDesiredCapabilities($test);
+        $requiredCapabilities = $resolver->resolveRequiredCapabilities($test);
+
+        $this->assertNotEmpty(
+            $desiredCapabilities->getCapability(CapabilitiesResolverFixture::CUSTOM_DESIRED_CAPABILITY)
+        );
+
+        $this->assertNotEmpty(
+            $requiredCapabilities->getCapability(CapabilitiesResolverFixture::CUSTOM_REQUIRED_CAPABILITY)
+        );
     }
 }